azure_server_client/requirements.txt
boto
boto3
botocore
certifi
click
cycler
dataclasses
Flask
Flask-Cors
importlib-metadata
itsdangerous
Jinja2
jmespath
MarkupSafe
numpy
pandas
pyparsing
python-dateutil
pytz
s3transfer
six
typing_extensions
urllib3
Werkzeug
zipp
azure-cosmos
aiohttp
